cleanup: refactor mocks to new namespace and directory
Move the mocks to `google::cloud::spanner_mocks::` (and
`google/cloud/spanner/mocks`). Also created a new library for the mocks,
which we will probably want to install.
