+++ cmake/CreateBazelConfig.cmake
@@ -58,16 +58,20 @@ function (create_bazel_config TARGET)
         get_target_property(sources ${TARGET} SOURCES)
     endif ()
     foreach (src ${sources})
+        # Some files need to be specificied with an absolute path (mainly
+        # sources for INTERFACE libraries). Compute the relative path because
+        # Bazel does not like absolute filenames.
+        string(REPLACE "${CMAKE_CURRENT_SOURCE_DIR}/" "" relative "${src}")
         string(FIND "${src}" "${CMAKE_CURRENT_BINARY_DIR}" in_binary_dir)
         if ("${in_binary_dir}" EQUAL 0)
             # Skip files in the binary directory, they are generated and handled
             # differently by our Bazel BUILD files.
         elseif("${src}" MATCHES "\\.inc$")
-            list(APPEND H ${src})
+            list(APPEND H ${relative})
         elseif("${src}" MATCHES "\\.h$")
-            list(APPEND H ${src})
+            list(APPEND H ${relative})
         elseif("${src}" MATCHES "\\.cc$")
-            list(APPEND CC ${src})
+            list(APPEND CC ${relative})
         endif ()
     endforeach ()
     write_bazel_copyright(${filename} ${_CREATE_BAZEL_CONFIG_OPT_YEAR})
