+++ google/cloud/spanner/BUILD
@@ -49,6 +49,20 @@ cc_library(
     ],
 )
 
+load(":spanner_client_mocks.bzl", "spanner_client_mocks_hdrs", "spanner_client_mocks_srcs")
+
+cc_library(
+    name = "spanner_client_mocks",
+    srcs = spanner_client_mocks_srcs,
+    hdrs = spanner_client_mocks_hdrs,
+    deps = [
+        ":spanner_client",
+        "@com_github_googleapis_google_cloud_cpp//google/cloud:google_cloud_cpp_common",
+        "@com_github_googleapis_google_cloud_cpp//google/cloud/testing_util:google_cloud_cpp_testing",
+        "@com_google_googletest//:gtest",
+    ],
+)
+
 load(":spanner_client_testing.bzl", "spanner_client_testing_hdrs", "spanner_client_testing_srcs")
 
 cc_library(
@@ -57,6 +71,7 @@ cc_library(
     hdrs = spanner_client_testing_hdrs,
     deps = [
         ":spanner_client",
+        ":spanner_client_mocks",
         "@com_github_googleapis_google_cloud_cpp//google/cloud:google_cloud_cpp_common",
         "@com_github_googleapis_google_cloud_cpp//google/cloud/testing_util:google_cloud_cpp_testing",
         "@com_google_googletest//:gtest",
@@ -70,6 +85,7 @@ load(":spanner_client_unit_tests.bzl", "spanner_client_unit_tests")
     srcs = [test],
     deps = [
         ":spanner_client",
+        ":spanner_client_mocks",
         ":spanner_client_testing",
         "@com_github_googleapis_google_cloud_cpp//google/cloud:google_cloud_cpp_common",
         "@com_github_googleapis_google_cloud_cpp//google/cloud/testing_util:google_cloud_cpp_testing",
