+++ google/cloud/spanner/mocks/mock_instance_admin_connection.h
@@ -12,15 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef GOOGLE_CLOUD_CPP_SPANNER_GOOGLE_CLOUD_SPANNER_TESTING_MOCK_INSTANCE_ADMIN_CONNECTION_H_
-#define GOOGLE_CLOUD_CPP_SPANNER_GOOGLE_CLOUD_SPANNER_TESTING_MOCK_INSTANCE_ADMIN_CONNECTION_H_
+#ifndef GOOGLE_CLOUD_CPP_SPANNER_GOOGLE_CLOUD_SPANNER_MOCKS_MOCK_INSTANCE_ADMIN_CONNECTION_H_
+#define GOOGLE_CLOUD_CPP_SPANNER_GOOGLE_CLOUD_SPANNER_MOCKS_MOCK_INSTANCE_ADMIN_CONNECTION_H_
 
 #include "google/cloud/spanner/instance_admin_connection.h"
 #include <gmock/gmock.h>
 
 namespace google {
 namespace cloud {
-namespace spanner_testing {
+namespace spanner_mocks {
 inline namespace SPANNER_CLIENT_NS {
 class MockInstanceAdminConnection
     : public google::cloud::spanner::InstanceAdminConnection {
@@ -31,8 +31,8 @@ class MockInstanceAdminConnection
 };
 
 }  // namespace SPANNER_CLIENT_NS
-}  // namespace spanner_testing
+}  // namespace spanner_mocks
 }  // namespace cloud
 }  // namespace google
 
-#endif  // GOOGLE_CLOUD_CPP_SPANNER_GOOGLE_CLOUD_SPANNER_TESTING_MOCK_INSTANCE_ADMIN_CONNECTION_H_
+#endif  // GOOGLE_CLOUD_CPP_SPANNER_GOOGLE_CLOUD_SPANNER_MOCKS_MOCK_INSTANCE_ADMIN_CONNECTION_H_
