+++ google/cloud/spanner/mocks/mock_instance_admin_stub.h
@@ -12,15 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef GOOGLE_CLOUD_CPP_SPANNER_GOOGLE_CLOUD_SPANNER_TESTING_MOCK_INSTANCE_ADMIN_STUB_H_
-#define GOOGLE_CLOUD_CPP_SPANNER_GOOGLE_CLOUD_SPANNER_TESTING_MOCK_INSTANCE_ADMIN_STUB_H_
+#ifndef GOOGLE_CLOUD_CPP_SPANNER_GOOGLE_CLOUD_SPANNER_MOCKS_MOCK_INSTANCE_ADMIN_STUB_H_
+#define GOOGLE_CLOUD_CPP_SPANNER_GOOGLE_CLOUD_SPANNER_MOCKS_MOCK_INSTANCE_ADMIN_STUB_H_
 
 #include "google/cloud/spanner/internal/instance_admin_stub.h"
 #include <gmock/gmock.h>
 
 namespace google {
 namespace cloud {
-namespace spanner_testing {
+namespace spanner_mocks {
 inline namespace SPANNER_CLIENT_NS {
 class MockInstanceAdminStub
     : public google::cloud::spanner::internal::InstanceAdminStub {
@@ -33,8 +33,8 @@ class MockInstanceAdminStub
 };
 
 }  // namespace SPANNER_CLIENT_NS
-}  // namespace spanner_testing
+}  // namespace spanner_mocks
 }  // namespace cloud
 }  // namespace google
 
-#endif  // GOOGLE_CLOUD_CPP_SPANNER_GOOGLE_CLOUD_SPANNER_TESTING_MOCK_INSTANCE_ADMIN_STUB_H_
+#endif  // GOOGLE_CLOUD_CPP_SPANNER_GOOGLE_CLOUD_SPANNER_MOCKS_MOCK_INSTANCE_ADMIN_STUB_H_
