+++ google/cloud/spanner/spanner_client_mocks.bzl
@@ -14,17 +14,15 @@
 #
 # DO NOT EDIT -- GENERATED BY CMake -- Change the CMakeLists.txt file if needed
 
-"""Automatically generated source lists for spanner_client_testing - DO NOT EDIT."""
+"""Automatically generated source lists for spanner_client_mocks - DO NOT EDIT."""
 
-spanner_client_testing_hdrs = [
-    "testing/database_environment.h",
-    "testing/matchers.h",
-    "testing/random_database_name.h",
-    "testing/validate_metadata.h",
+spanner_client_mocks_hdrs = [
+    "mocks/mock_database_admin_stub.h",
+    "mocks/mock_instance_admin_connection.h",
+    "mocks/mock_instance_admin_stub.h",
+    "mocks/mock_spanner_connection.h",
+    "mocks/mock_spanner_stub.h",
 ]
 
-spanner_client_testing_srcs = [
-    "testing/database_environment.cc",
-    "testing/random_database_name.cc",
-    "testing/validate_metadata.cc",
+spanner_client_mocks_srcs = [
 ]
